# docs/swagger.yaml
definitions:
...
  requests.CreateReviewRequest:
    properties:
      comment:
        type: string
      product_id:
        type: integer
      rating:
        type: integer
      user_id:
        type: integer
    type: object
...
  requests.UpdateReviewRequest:
    properties:
      comment:
        type: string
      name:
        type: string
      rating:
        type: integer
      review_id:
        type: integer
    required:
    - comment
    - name
    - rating
    - review_id
    type: object
...
  response.ApiResponsePaginationReview:
    properties:
      data:
        items:
          $ref: '#/definitions/response.ReviewResponse'
        type: array
      message:
        type: string
      pagination:
        $ref: '#/definitions/response.PaginationMeta'
      status:
        type: string
    type: object
  response.ApiResponsePaginationReviewDeleteAt:
    properties:
      data:
        items:
          $ref: '#/definitions/response.ReviewResponseDeleteAt'
        type: array
      message:
        type: string
      pagination:
        $ref: '#/definitions/response.PaginationMeta'
      status:
        type: string
    type: object
...
  response.ApiResponseProduct:
    properties:
      data:
        $ref: '#/definitions/response.ProductResponse'
      message:
        type: string
      status:
        type: string
    type: object
...
  response.ApiResponseReview:
    properties:
      data:
        $ref: '#/definitions/response.ReviewResponse'
      message:
        type: string
      status:
        type: string
    type: object
  response.ApiResponseReviewAll:
    properties:
      message:
        type: string
      status:
        type: string
    type: object
  response.ApiResponseReviewDelete:
    properties:
      message:
        type: string
      status:
        type: string
    type: object
  response.ApiResponseReviewDeleteAt:
    properties:
      data:
        $ref: '#/definitions/response.ReviewResponseDeleteAt'
      message:
        type: string
      status:
        type: string
    type: object
...
  response.ErrorResponse:
    properties:
      code:
        type: integer
      message:
        type: string
      status:
        type: string
    type: object
...
  response.PaginationMeta:
    properties:
      current_page:
        type: integer
      page_size:
        type: integer
      total_pages:
        type: integer
      total_records:
        type: integer
    type: object
  response.ProductResponse:
    properties:
      brand:
        type: string
      category_id:
        type: integer
      count_in_stock:
        type: integer
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      image_product:
        type: string
      merchant_id:
        type: integer
      name:
        type: string
      price:
        type: integer
      rating:
        type: number
      slug_product:
        type: string
      updated_at:
        type: string
      weight:
        type: integer
    type: object
...
  response.ReviewResponse:
    properties:
      comment:
        type: string
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      product_id:
        type: integer
      rating:
        type: integer
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
  response.ReviewResponseDeleteAt:
    properties:
      comment:
        type: string
      created_at:
        type: string
      deleted_at:
        type: string
      id:
        type: integer
      name:
        type: string
      product_id:
        type: integer
      rating:
        type: integer
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
...
info:
  contact: {}
paths:
...
  /api/product/{id}:
    get:
      consumes:
      - application/json
      description: Retrieve a product by ID
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Product data
          schema:
            $ref: '#/definitions/response.ApiResponseProduct'
        "400":
          description: Invalid product ID
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Failed to retrieve product data
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: Find product by ID
      tags:
      - Product
...
  /api/review:
    get:
      consumes:
      - application/json
      description: Retrieve a list of all review
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Number of items per page
        in: query
        name: page_size
        type: integer
      - description: Search query
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of review
          schema:
            $ref: '#/definitions/response.ApiResponsePaginationReview'
        "500":
          description: Failed to retrieve review data
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: Find all review
      tags:
      - Review
...
  /api/review/active:
    get:
      consumes:
      - application/json
      description: Retrieve a list of active review
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Number of items per page
        in: query
        name: page_size
        type: integer
      - description: Search query
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of active review
          schema:
            $ref: '#/definitions/response.ApiResponsePaginationReviewDeleteAt'
        "500":
          description: Failed to retrieve review data
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: Retrieve active review
      tags:
      - Review
  /api/review/create:
    post:
      consumes:
      - application/json
      description: Create a new review with the provided details
      parameters:
      - description: review details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/requests.CreateReviewRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Successfully created review
          schema:
            $ref: '#/definitions/response.ApiResponseReview'
        "400":
          description: Invalid request body or validation error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Failed to create review
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: Create a new review
      tags:
      - Review
  /api/review/delete/{id}:
    delete:
      consumes:
      - application/json
      description: Permanently delete a review record by its ID.
      parameters:
      - description: review ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully deleted review record permanently
          schema:
            $ref: '#/definitions/response.ApiResponseReviewDelete'
        "400":
          description: 'Bad Request: Invalid ID'
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: 'Failed to delete review:'
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: Permanently delete a review
      tags:
      - review
  /api/review/delete/all:
    post:
      consumes:
      - application/json
      description: Permanently delete a review record by its ID.
      parameters:
      - description: review ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully deleted review record permanently
          schema:
            $ref: '#/definitions/response.ApiResponseReviewAll'
        "400":
          description: 'Bad Request: Invalid ID'
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: 'Failed to delete review:'
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: Permanently delete a review
      tags:
      - Review
  /api/review/merchant/{id}:
    get:
      consumes:
      - application/json
      description: Retrieve a list of reviews for a specific merchant
      parameters:
      - description: merchant ID
        in: path
        name: id
        required: true
        type: integer
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Number of items per page
        in: query
        name: page_size
        type: integer
      - description: Search query
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of reviews for the merchant
          schema:
            $ref: '#/definitions/response.ApiResponsePaginationReview'
        "400":
          description: Invalid merchant ID
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Failed to retrieve review data
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: Find reviews by merchant ID
      tags:
      - Review
  /api/review/product/{id}:
    get:
      consumes:
      - application/json
      description: Retrieve a list of reviews for a specific product
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Number of items per page
        in: query
        name: page_size
        type: integer
      - description: Search query
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of reviews for the product
          schema:
            $ref: '#/definitions/response.ApiResponsePaginationReview'
        "400":
          description: Invalid product ID
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Failed to retrieve review data
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: Find reviews by product ID
      tags:
      - Review
  /api/review/restore/{id}:
    post:
      consumes:
      - application/json
      description: Restore a trashed review record by its ID.
      parameters:
      - description: Review ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully restored review
          schema:
            $ref: '#/definitions/response.ApiResponseReviewDeleteAt'
        "400":
          description: Invalid review ID
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Failed to restore review
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: Restore a trashed review
      tags:
      - Review
  /api/review/restore/all:
    post:
      consumes:
      - application/json
      description: Restore a trashed review record by its ID.
      parameters:
      - description: review ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully restored review all
          schema:
            $ref: '#/definitions/response.ApiResponseReviewAll'
        "400":
          description: Invalid review ID
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Failed to restore review
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: Restore a trashed review
      tags:
      - review
  /api/review/trashed:
    get:
      consumes:
      - application/json
      description: Retrieve a list of trashed review records
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Number of items per page
        in: query
        name: page_size
        type: integer
      - description: Search query
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of trashed review data
          schema:
            $ref: '#/definitions/response.ApiResponsePaginationReviewDeleteAt'
        "500":
          description: Failed to retrieve review data
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: Retrieve trashed review
      tags:
      - Review
  /api/review/trashed/{id}:
    get:
      consumes:
      - application/json
      description: Retrieve a trashed review record by its ID.
      parameters:
      - description: review ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved trashed review
          schema:
            $ref: '#/definitions/response.ApiResponseReviewDeleteAt'
        "400":
          description: Invalid request body or validation error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Failed to retrieve trashed review
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: Retrieve a trashed review
      tags:
      - Review
  /api/review/update/{id}:
    post:
      consumes:
      - application/json
      description: Update an existing review record with the provided details
      parameters:
      - description: review ID
        in: path
        name: id
        required: true
        type: integer
      - description: review update details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/requests.UpdateReviewRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully updated review
          schema:
            $ref: '#/definitions/response.ApiResponseReview'
        "400":
          description: Invalid request body or validation error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Failed to update review
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: Update an existing review
      tags:
      - Review
...
swagger: "2.0"